Added in Enemy Spawner and A Collectable Interruption Spawner. Fixed the Cmd Prompt

+++ Community Game Jam 2k19/Assets/Scripts/InterruptionObject.cs
@@ -6,5 +6,5 @@ using UnityEngine;
 public class InterruptionObject : ScriptableObject
 {
     public float speed;
-    public float difficulty;
+    public int difficulty;
 }
